make the audio-capture speech-to-text sample work
Update the server speech handler - it now creates a second temporary
file with the audio in it, because the one that the server framework
creates doesn't have an extension, and codekit (ruby) relies on the
extension to determine the MIME type of the audio file.

Modify the result display in the speech app2 sample, so it actually
displays.

Tweak the jquery ajax call that sends audio up to the server, so it
doesn't try to modify the data before sending it.

Modify codekit (ruby) to open audio files in binary mode, so the
contents don't get truncated.

modify recorderjs to encode WAV files in mono - the AT&T speech-to-text
can't handle stereo input.